+++ src/modify/cvolton.betterinfo/RewardViewLayer.cpp
@@ -1,20 +1,16 @@
 #include <Geode/Geode.hpp>
 #include "../../SwelvyBG.hpp"
 #include "../../Hooks/Hooker.hpp" 
-class RewardViewLayer : public Betterhook::HookBetter { 
-    void init(CCNode* _This) override {
-         if (auto bg = _This->getChildByID("cvolton.betterinfo/background")) {
+
+using namespace geode::prelude;
+Viper_Hookclass(RewardViewLayer) {
+        if (auto bg = this->getChildByID("cvolton.betterinfo/background")) {
             bg->setVisible(false);
         }
         SwelvyBG* swelvyBG = SwelvyBG::create();
         swelvyBG->setZOrder(-1);
         swelvyBG->setID("swelvy-background");
-        _This->addChild(swelvyBG);
-    }
+        this->addChild(swelvyBG);
+}
 
-    const char* PutLayer() const override { return "RewardViewLayer"; }
-};
-
-
-REGISTER_HookBetter(RewardViewLayer);