+++ src/modify/minemaker0430.gddp_integration/DPLayer.cpp
@@ -1,21 +1,16 @@
 #include <Geode/Geode.hpp>
 #include "../../SwelvyBG.hpp"
 #include "../../Hooks/Hooker.hpp" 
-class DPLayer : public Betterhook::HookBetter { 
-    void init(CCNode* _This) override {
-         if (auto bg = _This->getChildByID("bg")) {
+
+
+using namespace geode::prelude;
+Viper_Hookclass(DPLayer) {
+        if (auto bg = this->getChildByID("bg")) {
             bg->setVisible(false);
         }
 
         SwelvyBG* swelvyBG = SwelvyBG::create();
         swelvyBG->setZOrder(-99999);
         swelvyBG->setID("swelvy-background");
-        _This->addChild(swelvyBG);
-    }
-
-    const char* PutLayer() const override { return "DPLayer"; }
-};
-
-
-REGISTER_HookBetter(DPLayer);
-
+        this->addChild(swelvyBG);
+}